Move all classes to internal package
