+++ addons/binding/org.openhab.binding.unifi/META-INF/MANIFEST.MF
@@ -7,8 +7,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-SymbolicName: org.openhab.binding.unifi;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.4.0.qualifier
-Export-Package: org.openhab.binding.unifi,
- org.openhab.binding.unifi.handler
 Import-Package: 
  com.google.gson,
  com.google.gson.annotations,
@@ -30,8 +28,6 @@ Import-Package:
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
  org.eclipse.smarthome.io.net.http,
- org.openhab.binding.unifi,
- org.openhab.binding.unifi.handler,
  org.osgi.framework,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
