+++ addons/binding/org.openhab.binding.unifi/src/main/java/org/openhab/binding/unifi/internal/UniFiBindingConstants.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.unifi;
+package org.openhab.binding.unifi.internal;
 
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
