+++ addons/binding/org.openhab.binding.unifi/src/main/java/org/openhab/binding/unifi/internal/UniFiThingHandlerFactory.java
@@ -20,8 +20,8 @@ import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.eclipse.smarthome.io.net.http.HttpClientFactory;
-import org.openhab.binding.unifi.handler.UniFiClientThingHandler;
-import org.openhab.binding.unifi.handler.UniFiControllerThingHandler;
+import org.openhab.binding.unifi.internal.handler.UniFiClientThingHandler;
+import org.openhab.binding.unifi.internal.handler.UniFiControllerThingHandler;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
 import org.osgi.service.component.annotations.Reference;
